requirements.txt
tornado
mysql-connector-python
sqlalchemy
sockjs-tornado
psutil
pyecharts
pyecharts-snapshot
